+++ src/Logger.ts
@@ -2,8 +2,6 @@ import randomColor from 'randomcolor';
 import Style from './types/Style';
 import ComponentLogger from './ComponentLogger';
 import { Options } from './types/Options';
-import BrowserLogger from './loggers/BrowserLogger';
-import ChalkLogger from './loggers/ChalkLogger';
 
 /**
  * The component background color
@@ -83,6 +81,5 @@ export default class Logger {
 	}
 }
 
-export { BrowserLogger, ChalkLogger };
-
 export * from './lib/cyclic-serializer';
+export * from './loggers/index';