+++ src/types/LogFunctionFactory.ts
@@ -1,0 +1,10 @@
+import LogFunction from './LogFunction';
+
+/**
+ * Builds a {@link LogFunction}
+ * @param options for the log function
+ * @return the {@link LogFunction}
+ */
+type LogFunctionFactory = (...args: any[]) => LogFunction;
+
+export default LogFunctionFactory;