+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Breaking changes
+- made `BrowserLogger` a `LogFunctionFactory` (`() => LogFunction`) for consistency.
+  
+  **Upgrade instructions:**
+  To upgrade, replace `BrowserLogger` with `BrowserLogger()` in your code.
 ### Added
 - `ChalkLogger` - a [chalk](https://github.com/chalk/chalk) based logger for NodeJS environments:
   ```ts